chore: remove duplicate expr editor
