--- frontend/src/components/DatabaseGroup/common/ExprEditor/components/index.ts
@@ -1,5 +1,0 @@
-import FactorSelect from "./FactorSelect.vue";
-import OperatorSelect from "./OperatorSelect.vue";
-import ValueInput from "./ValueInput.vue";
-
-export { FactorSelect, OperatorSelect, ValueInput };
