--- frontend/src/components/DatabaseGroup/common/ExprEditor/factor.ts
@@ -1,11 +1,0 @@
-export const StringFactorList = [
-  "resource.database_name",
-  "resource.schema_name",
-  "resource.table_name",
-  "resource.instance_id",
-] as const;
-
-export const FactorList = {
-  DATABASE_GROUP: ["resource.database_name", "resource.instance_id"],
-  SCHEMA_GROUP: ["resource.table_name"],
-};
