--- frontend/src/components/DatabaseGroup/common/ExprEditor/index.ts
@@ -1,3 +1,0 @@
-import ExprEditor from "./ExprEditor.vue";
-
-export default ExprEditor;
