+++ frontend/src/locales/en-US.json
@@ -3269,11 +3269,6 @@
       "create": "New table group",
       "edit": "Edit table group"
     },
-    "factor": {
-      "resource_instance_id": "Instance ID",
-      "resource_database_name": "Database name",
-      "resource_table_name": "Table name"
-    },
     "condition": {
       "self": "Condition"
     },
@@ -3339,7 +3334,10 @@
       "export_rows": "Export rows",
       "table_name": "Table name",
       "column_name": "Column name",
-      "classification_level": "Classification level"
+      "classification_level": "Classification level",
+      "resource_instance_id": "Instance ID",
+      "resource_database_name": "Database name",
+      "resource_table_name": "Table name"
     }
   },
   "changelist": {
