+++ frontend/src/locales/es-ES.json
@@ -3266,11 +3266,6 @@
       "create": "Crear grupo de tablas",
       "edit": "Editar grupo de tablas"
     },
-    "factor": {
-      "resource_instance_id": "ID de instancia",
-      "resource_database_name": "Nombre de la base de datos",
-      "resource_table_name": "Nombre de la tabla"
-    },
     "condition": {
       "self": "Condición"
     },
@@ -3337,7 +3332,10 @@
       "export_rows": "Exportar filas",
       "table_name": "Nombre de la tabla",
       "column_name": "Nombre de columna",
-      "classification_level": "Nivel de clasificación"
+      "classification_level": "Nivel de clasificación",
+      "resource_instance_id": "ID de instancia",
+      "resource_database_name": "Nombre de la base de datos",
+      "resource_table_name": "Nombre de la tabla"
     }
   },
   "changelist": {
