+++ frontend/src/locales/zh-CN.json
@@ -3262,11 +3262,6 @@
       "create": "新建表分组",
       "edit": "编辑表分组"
     },
-    "factor": {
-      "resource_instance_id": "实例 ID",
-      "resource_database_name": "数据库名",
-      "resource_table_name": "表名"
-    },
     "condition": {
       "self": "条件"
     },
@@ -3332,7 +3327,10 @@
       "export_rows": "导出行数",
       "table_name": "表名称",
       "column_name": "字段名称",
-      "classification_level": "数据分级"
+      "classification_level": "数据分级",
+      "resource_instance_id": "实例 ID",
+      "resource_database_name": "数据库名",
+      "resource_table_name": "表名"
     }
   },
   "changelist": {
